jewelry-mvp-bracelet/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pillow>=10.0.0
colorthief>=0.2.1
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.26.0
numpy>=1.26.0
google-generativeai>=0.3.0
gradio_client>=1.0.0
trimesh>=4.0.0
